pyproject.toml
[tool.poetry]
name = "python-polish-version"
version = "1.0.0"
description = "Translator from polish version of Python to Python 3.13"
authors = ["Michał Proć <michalproc4@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
antlr4-python3-runtime = "^4.13"
pyyml = "^0.0.2"
flake8 = "^7.1.1"
flake8-json = "^24.4.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
